// src/models/review.ts
export type UpdateReviewDto = {
  reviewId: string;
  reviewText: string;
  userRating: number;
};

export type CreateReviewDto = {
  hotelroomid: string;
  reviewText: string;
  userRating: number;
  userId: string;
};

export type Review = {
  text: string;
  user: { name: string };
  userRating: number;
  _createdAt: Date;
  _id: string;
};